Small changes to `excel2csv`
