--- .github/workflows/excel2csv.yml
@@ -1,32 +1,0 @@
-name: excel2csv
-
-on: [push]
-
-jobs:
-  run:
-    name: excel2csv
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v3
-
-      - name: Set up Python 3.10
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.10'
-          cache: 'pip'
-          cache-dependency-path: 'pyqalloy-contribute/requirements.txt'
-
-      - name: Install Dependencies
-        run: |
-          python -m pip install -r pyqalloy-contribute/requirements.txt
-
-      - name: Run excel2csv
-        run: |
-          python pyqalloy-contribute/pyqalloy-contribute/excel2csv.py .
-
-      - name: Commit changes with Add & Commit
-        uses: EndBug/add-and-commit@v9
-        with:
-          message: '(automatic) excel2csv Action for Data Tracking'
-          add: '*.csv'
